requirements.txt
fastapi
pandas
plotly
psycopg2-binary
uvicorn